Refactor login flow with updated backend
