+++ src/modules/auth/LoginForm/LoginForm.tsx
@@ -14,16 +14,18 @@ import {
   Link,
   FormErrorMessage,
 } from '@common/components';
+import { LoginError } from '@generated/graphql.queries';
+import { isNil } from 'lodash';
 
 type LoginFormProps = {
   onSubmit: (values: LoginFormTypes) => void;
   isLoading: boolean;
-  isInvalid: boolean;
+  error: LoginError | null;
 };
 export default function LoginForm({
   onSubmit,
   isLoading,
-  isInvalid,
+  error,
 }: LoginFormProps) {
   const { t } = useTranslation('auth');
 
@@ -72,7 +74,7 @@ export default function LoginForm({
       maxW="48rem"
     >
       <form onSubmit={handleSubmit(onSubmit)}>
-        <FormControl isInvalid={isInvalid}>
+        <FormControl isInvalid={!isNil(error)}>
           <VStack align="stretch" gap="3rem">
             {formValues.inputs.map(
               ({ type, name, placeholder, leftElement }, index) => (
@@ -86,7 +88,10 @@ export default function LoginForm({
                 />
               ),
             )}
-            <FormErrorMessage testId="login.error" error={t('login.error')} />
+            <FormErrorMessage
+              testId="login.error"
+              error={t(`login.error.${error?.key}`)}
+            />
             <Button
               type="submit"
               isLoading={isLoading}
