--- src/modules/auth/LoginForm/login.gql
@@ -1,11 +1,0 @@
-mutation Login($payload: LoginDetailInput!) {
-  login(detail: $payload) {
-    accessToken
-    user {
-      id
-      status
-      username
-      email
-    }
-  }
-}
